+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- CSS Handles to the app
+- CSS Handles to the app.
 
 ### Fixed
-- Update docs
+- Update docs.
+
+### Remove
+- Not implemented props `showPromotionCategory` and `showGiftCategory`.
 
 ## [2.13.2] - 2019-08-29
 ### Changed